Change the replica number to be 3.

+++ test/edu/ucsb/nceas/metacat/dataone/MNodeReplicationTest.java
@@ -199,7 +199,7 @@ public class MNodeReplicationTest extends D1NodeServiceTest {
         }
         ReplicationPolicy rePolicy = new ReplicationPolicy();
         rePolicy.setReplicationAllowed(true);
-        rePolicy.setNumberReplicas(new Integer(2));
+        rePolicy.setNumberReplicas(new Integer(3));
         rePolicy.addPreferredMemberNode(localNode.getIdentifier());
         sysmeta.setReplicationPolicy(rePolicy);
